requirements.txt
Flask==0.10.1
Flask-SocketIO==2.6
python-socketio==1.4.2
PyYAML==3.11
Jinja2==2.8
WTForms==2.1
gevent-websocket==0.9.5
Werkzeug==0.11.4
python-dateutil==2.5.3
